+++ src/bin/cmd/random.rs
@@ -1,30 +1,18 @@
 use std::io::{self, ErrorKind};
-use clap::ArgMatches;
 use famlib::{random::RandomGen};
 use crate::data::{DataSink, DataSource};
-use super::Command;
+use super::{Command, datasink, datasource};
 
 pub struct Random{}
 
 impl Random {
-    fn _get_input(matches: &ArgMatches) -> DataSource {
-        match matches.value_of("input") {
-            None => DataSource::stdin(),
-            Some(x) => DataSource::from(&x),
-        }
-    }
-    fn _get_output(matches: &ArgMatches) -> DataSink {
-        match matches.value_of("output") {
-            None => DataSink::StdOut,
-            Some(x) => DataSink::FilePath(String::from(x)),
-        }
-    }
-    pub fn shuffle_command(fs: DataSource, fo: DataSink, fixed: bool) -> io::Result<()> {
+    pub fn shuffle_command(fs: DataSource, fo: DataSink, fixed: bool)
+            -> io::Result<()> {
         let msa = fs.get_sequence_collection().unwrap().to_msa();
         match msa {
             Ok(mut msa) => {
                 msa.shuffle(fixed);
-                fo.write_fasta(msa)
+                fo.write_fasta(&msa)
             }
             Err(_) => {
                 Err(std::io::Error::new(
@@ -33,12 +21,13 @@ impl Random {
                 }
             }
     }
-    pub fn shuffle_rows_command(fs: DataSource, fo: DataSink) -> io::Result<()> {
+    pub fn shuffle_rows_command(fs: DataSource, fo: DataSink)
+            -> io::Result<()> {
         let msa = fs.get_sequence_collection().unwrap().to_msa();
         match msa {
             Ok(mut msa) => {
                 msa.shuffle_rows();
-                fo.write_fasta(msa)
+                fo.write_fasta(&msa)
             }
             Err(_) => {
                 Err(std::io::Error::new(
@@ -48,12 +37,13 @@ impl Random {
         }
 
     }
-    pub fn shuffle_cols_command(fs: DataSource, fo: DataSink) -> io::Result<()> {
+    pub fn shuffle_cols_command(fs: DataSource, fo: DataSink)
+            -> io::Result<()> {
         let msa = fs.get_sequence_collection().unwrap().to_msa();
         match msa {
             Ok(mut msa) => {
                 msa.shuffle_rows();
-                fo.write_fasta(msa)
+                fo.write_fasta(&msa)
             }
             Err(_) => {
                 Err(std::io::Error::new(
@@ -72,19 +62,19 @@ impl Command for Random {
                     let fixed = sm.map_or_else(
                         || false, |x| x.is_present("fixed"));
                     Self::shuffle_command(
-                        Self::_get_input(m),
-                        Self::_get_output(m),
+                        datasource(m),
+                        datasink(m),
                         fixed)?
                 },
                 ("rows", _) => {
                     Self::shuffle_rows_command(
-                        Self::_get_input(m),
-                        Self::_get_output(m))?
+                        datasource(m),
+                        datasink(m))?
                 },
                 ("cols", _) => {
                     Self::shuffle_cols_command(
-                        Self::_get_input(m),
-                        Self::_get_output(m))?
+                        datasource(m),
+                        datasink(m))?
                 },
                 (_, _) => {}
             };